+++ src/App.vue
@@ -40,7 +40,6 @@ export default {
     firebase.auth().onAuthStateChanged(user => {
       if (user) {
         this.setLoginUser(user);
-        this.fetchLoginUsers();
         this.fetchWork();
         this.fetchProfile();
         this.fetchTime();
@@ -86,7 +85,6 @@ export default {
     ...mapActions([
       "toggleSideMenu",
       "setLoginUser",
-      "fetchLoginUsers",
       "deleteLoginUser",
       "fetchWork",
       "fetchProfile",
